app/requirements.txt
flask
mysql.connector
cryptography
Flask-MySQL
PyMySQL
mysql-connector-python